2025 해커톤/virtual_virtual_streamer/requirement.txt
fastapi
uvicorn[standard]
pydantic>=2
sqlalchemy>=2
python-dotenv
google-generativeai>=0.5     
openai-whisper
sounddevice                 
coqui-tts
python-multipart             # 음성 업로드용
pytchat                      # 유튜브 라이브 채팅 수집
asyncio-throttle             # 비동기 처리 제한